AgriBuddy/AgriBuddy/requirements.txt
# requirements.txt — Databricks Apps runtime
# Switched from Gradio to FastAPI to avoid huggingface_hub version conflicts.
# FastAPI + uvicorn have zero dependency conflicts in Databricks Apps.

fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.30.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
prophet>=1.1.5
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
pymupdf>=1.24.0
mlflow>=2.13.0
databricks-sdk>=0.28.0
python-dotenv>=1.0.0
databricks-sql-connector>=3.0.0


